this commit replaces the previous dynamic config logic with a schema builder which builds a schema of the config using reflection this allows us to more robustly handle config elements, set/get/check them etc. + need more tests and then we can join the ast tree with the schema to finally parse out a config

+++ internal/cluster/gbNode.go
@@ -388,7 +388,7 @@ func (s *GBServer) buildAddrGroupMap(known []string) (map[string][]string, error
 			continue
 		}
 
-		if sizeEstimate+len(name) > DEFAULT_MAX_DISCOVERY_SIZE {
+		if uint16(sizeEstimate+len(name)) > DEFAULT_MAX_DISCOVERY_SIZE {
 			return addrMap, nil
 		}
 
@@ -400,7 +400,7 @@ func (s *GBServer) buildAddrGroupMap(known []string) (map[string][]string, error
 			if value.KeyGroup == ADDR_DKG {
 				log.Printf("tcpKey = %v", key)
 
-				if sizeEstimate+len(key) > DEFAULT_MAX_DISCOVERY_SIZE {
+				if uint16(sizeEstimate+len(key)) > DEFAULT_MAX_DISCOVERY_SIZE {
 					return addrMap, nil
 				} else {
 					addrMap[name] = append(addrMap[name], value.Key)
